API

This section documents all functions not documented elsewhere

TheGraphOpt.x — Function

x(g::GradientDescent)
x(g::GradientDescent, v)

The current best guess for the solution. If using the setter, v is the new value.

The setter is not in-place. See TheGraphOpt.x!.

TheGraphOpt.x! — Function

x!(g::GradientDescent, v)

In-place setting of g.x to v

See TheGraphOpt.x.

TheGraphOpt.η — Function

η(g::GradientDescent)
η(g::GradientDescent, v)

The learning rate/step size. If using the setter, v is the new value.

The setter is not in-place.

TheGraphOpt.ϵ — Function

ϵ(g::GradientDescent)
ϵ(g::GradientDescent, v)

The tolerance. If using the setter, v is the new value.

The setter is not in-place.

TheGraphOpt.maybeminimize! — Function

maybeminimize!(f::Function, a::OptAlgorithm, op::Function)

Minimize f using a, which calls op for updating a.x.

This function may be in-place. Don't use it directly unless you know what you're doing. This function is unexported.

Warning

If you don't provide any hook with the IsStoppingCondition trait, this will loop forever.

TheGraphOpt.shouldstop — Function

shouldstop(hs::AbstractVecOrTuple{H}, a::OptAlgorithm; locals...)

Map over the hooks hs to check for stopping conditions.

This performs an OR operation if there are multiple hooks that have the TheGraphOpt.IsStoppingCondition trait.

TheGraphOpt.iteration — Function

iteration(f::Function, a::OptAlgorithm)

One iteration of a on the function f.

This function is unexported.