Hooks

Hooks enable you to dynamically choose to inject functionality into algorithms. Some hooks are mandatory for certain algorithms, such as having a hook with the TheGraphOpt.IsStoppingCondition trait when using TheGraphOpt.GradientDescent. Others are purely there for you to use at your discretion. In this section, we'll take you through Traits, Using Predefined Hooks and how to Create Custom Hooks.

Traits

Hooks have traits. This is how our code knows when to execute which hook. For example, the code will execute a hook that has the TheGraphOpt.IsStoppingCondition trait when evaluating whether it has finished optimising. A full list of traits follows.

StopTrait

This trait tells the code if it should execute a hook when checking stopping conditions. By default, the code automatically gives all hooks the TheGraphOpt.IsNotStoppingCondition trait except in certain situations, which are explicitly documented with the hooks in question. We decide to stop code execution on an OR basis. This means that if any hook with TheGraphOpt.IsStoppingCondition returns true, the code breaks out of the optimisation loop. Hooks with this trait must implement TheGraphOpt.stophook.

TheGraphOpt.StopTrait — Type

Abstract type for stopping conditions traits

TheGraphOpt.IsStoppingCondition — Type

Exhibited by hooks that are stopping conditions. Stopping condition hooks must emit a boolean value when TheGraphOpt.stophook is called. If multiple hooks meet IsStoppingCondition are instantiated at the same time, we assume that they are meant to be OR'ed, so if any of them is true, optimisation is finished. For more complex behaviour, consider defining a more complex function using a single StopWhen.

To set this trait for a hook, run

julia> StopTrait(::Type{MyHook}) = IsStoppingCondition()

TheGraphOpt.IsNotStoppingCondition — Type

Exhibited by hooks that are not stopping conditions. This is the default case. You should not have to set it manually for any hooks.

TheGraphOpt.stophook — Function

stophook(h::::IsStoppingCondition, a::OptAlgorithm; locals...)

Raise an error if the hook is a stopping condition but has not implemented stophook.

stophook(h::::(!IsStoppingCondition), a::OptAlgorithm; locals...)

If the hook isn't a stopping condition, it shouldn't be considered in the OR, so return false.

stophook(h::StopWhen, a::OptAlgorithm; locals...)

Call h.f on a and ;locals.

Using Predefined Hooks

Hooks always descend from the same abstract type.

TheGraphOpt.Hook — Type

Abstract type for hooks.

Generally speaking, pre-defined hooks won't exhibit any positive traits unless explicitly documented otherwise. This serves two purposes. Firstly, it prevents any unexpected behaviour when the code executes. Secondly, it gives you more flexibility when choosing where you want a hook to be executed.

Warning

This section is incomplete! We will fill this out more once we have more traits and hooks to work with.

Note

We recommend you read the below StopWhen section as it explains details that we won't cover in the later sections since it'd be too repetitive.

StopWhen

This hook has the TheGraphOpt.IsStoppingCondition trait. To use it, you would specify a function that returns a boolean value. If said value is true, then the code breaks out of the optimisation loop.

Let's take an example from our tests to demonstrate how to specify this hook. We'll also implement a dummy optimisation function that just implements a counter for illustration purposes.

julia> using TheGraphOpt
julia> struct FakeOptAlg <: TheGraphOpt.OptAlgorithm end
juila> a = FakeOptAlg()
julia> function counter(h, a)
            i = 0
            while !shouldstop(h, a; Base.@locals()...)
                i += 1
            end
            return i
        end
julia> h = StopWhen((a; locals...) -> locals[:i] ≥ 5, Dict())  # Stop when i ≥ 5
julia> i = counter((h,), a)
5

One thing you may not have seen is Base.@locals. This takes variables from the local scope (in this case, from the counter scope), and tracks them as a dictionary of symbols. Thus, since i is a local variable inside of counter, :i becomes a key in the Base.@locals dictionary. We pass this dictionary to the anonymouse function stored by StopWhen. Then, we can use locals[:i] to get the value of i from the counter scope and check it against some condition. This is a powerful trick you may find yourself using a lot when dealing with hooks.

TheGraphOpt.StopWhen — Type

StopWhen(f::Function)

Stops optimisation when some condition is met.

The condition is set by f. Note that f gets access to variables in the TheGraphOpt.minimize scope. This means, for example, that it can use locals[:z] to compute residuals. This has the TheGraphOpt.IsStoppingCondition trait.

Create Custom Hooks

When you create a custom hook, you need to follow three steps. The first is that you need to descend from TheGraphOpt.Hook.

julia> using TheGraphOpt
julia> struct MyHook <: Hook end

The second is that you need to ensure that you specify which traits you want that hook to exhibit. For example, let's say MyHook is a stopping condition. You'd want to implement.

julia> StopTrait(::Type{MyHook}) = IsStoppingCondition()

Finally, if you need non-default behaviour for when the hook executes, you'll need to implement whatever function(s) the code calls for that trait-type. For IsStoppingCondition, that's stophook. Say we want MyHook to immediately cause optimisation to finish. We'd implement

julia> stophook(h::MyHook, a::TheGraphOpt.OptAlgorithm; locals...) = true

That's it! As long as your follow those three steps, you should be able to implement whatever hook you want!