Algorithms

This section will introduce you to the basic workflow that you'll use to optimise functions, as well as the algorithms we've implemented.

Generally speaking, there are two parts to optimising. The first is to specify the function you want to optimise. Say,

julia> f(x) = sum(x .^ 2)

Then you want to create a Hook with the TheGraphOpt.IsStoppingCondition trait. Else, optimisation will get stuck in an infinite loop.

julia> using LinearAlgebra
julia> h = StopWhen((a; locals...) -> norm(x(a) - locals[:z]) < ϵ(a))  # Stop when the residual is less than the tolerance

Then, you want to choose the algorithm you want to use for minimisation and specify its parameters.

julia> a = GradientDescent(; x=[100.0, 50.0], η=1e-1, ϵ=1e-6, hooks=[h,])  # Specify parameters for optimisation

Finally, you'll run minimize! or minimize.

julia> sol = minimize!(f, a)  # Optimise

TheGraphOpt.minimize! — Function

minimize!(f::Function, a::OptAlgorithm)

Minimize f using a.

Does in-place updates of a.x. This will generally be more performant than TheGraphOpt.minimize. However, there are cases in which this will be worse, so we provide both.

Warning

If you don't provide any hook with the IsStoppingCondition trait, this will loop forever.

TheGraphOpt.minimize — Function

minimize(f::Function, a::OptAlgorithm)

Minimize f using a.

This will generally be less performant than TheGraphOpt.minimize!. However, there are cases in which this will be better, so we provide it as an option.

Warning

If you don't provide any hook with the IsStoppingCondition trait, this will loop forever.

sol here is a struct containing various metadata. If you only care about the optimal value of x, then grab it using

julia> TheGraphOpt.x(sol)

Algorithms

Gradient Descent

From Wikipedia:

[Gradient Descent]... is a first-order iterative optimization algorithm for finding a local minimum of a differentiable function. The idea is to take repeated steps in the opposite direction of the gradient (or approximate gradient) of the function at the current point, because this is the direction of steepest descent.

If the function we're optimising f is convex, then a local minimum is also a global minimum. The update rule for gradient descent is $x_{n+1}=x_n - η∇f(x_n)$.

TheGraphOpt.GradientDescent — Type

GradientDescent{
    T<:Real,H<:Hook,V<:AbstractVector{T},S<:AbstractVecOrTuple{H}
} <: OptAlgorithm

Specifies parameters for gradient descent learning.

Fields

η::T is the learning rate/step size.

x::V is the current best guess for the solution.

ϵ::T is the tolerance.

hooks::S are the hooks