API

This section documents all functions not documented elsewhere

TheGraphOpt.iteration — Function

iteration(f::Function, a::GradientDescent)

One iteration of a on the function f.

This function is unexported.

iteration(f::Function, a::ProjectedGradientDescent)

Apply a to f, projected based on a.t.

TheGraphOpt.f — Function

Getter for the stop-function.

TheGraphOpt.η — Function

η(g::GradientDescent)

The learning rate/step size.

TheGraphOpt.hooks — Function

hooks(g::GradientDescent)

The hooks used by the algorithm.

TheGraphOpt.λ — Function

Get the λ of h.

TheGraphOpt.maybeminimize! — Function

maybeminimize!(f::Function, a::OptAlgorithm, op::Function)

Minimize f using a, which calls op for updating a.x.

This function may be in-place. Don't use it directly unless you know what you're doing. This function is unexported.

Warning

If you don't provide any hook with the IsStoppingCondition trait, this will loop forever.

TheGraphOpt.postiteration — Function

postiterationhook(
    hs::AbstractVecOrTuple{H}, a::OptAlgorithm, z::AbstractVector{T}; locals...
) where {H<:Hook,T<:Real}

Run hooks that the code should execute after TheGraphOpt.iteration.

TheGraphOpt.shouldstop — Function

shouldstop(hs::AbstractVecOrTuple{H}, a::OptAlgorithm; locals...)

Map over the hooks hs to check for stopping conditions.

This performs an OR operation if there are multiple hooks that have the TheGraphOpt.IsStoppingCondition trait.

TheGraphOpt.x — Function

x(g::GradientDescent)
x(g::GradientDescent, v)

The current best guess for the solution. If using the setter, v is the new value.

The setter is not in-place. See TheGraphOpt.x!.

x(g::ProjectedGradientDescent)
x(g::ProjectedGradientDescent, v)

The current best guess for the solution. If using the setter, v is the new value.

The setter is not in-place. See TheGraphOpt.x!.

TheGraphOpt.x! — Function

x!(g::GradientDescent, v)

In-place setting of g.x to v

See TheGraphOpt.x.

x!(g::ProjectedGradientDescent, v)

In-place setting of a.g.x to v

See TheGraphOpt.x.

TheGraphOpt.x₀ — Function

Get the anchor of h.